add rabbit mq

+++ src/main/java/com/brash/service/MarkService.java
@@ -1,14 +1,14 @@
 package com.brash.service;
 
 import com.brash.data.entity.Mark;
-import com.brash.dto.MarkDTO;
+import com.brash.dto.web.MarkDTO;
 import com.brash.exception.NoAvailableMarkException;
 
 import java.util.List;
 
 public interface MarkService {
 
-    Mark addMark(int mark, long userId, long itemId);
+    Mark addMark(double mark, long userId, long itemId);
 
     Mark getMark(long userId, long itemId) throws NoAvailableMarkException;
     List<Mark> getGeneratedMarks(long userId) throws NoAvailableMarkException;
